pyproject.toml
[project]
name = "document-intelligence-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-documentintelligence>=1.0.0b4",
    "azure-identity>=1.19.0",
    "python-dotenv>=1.0.1",
]
